Use estimated file size for transcoded streams
Before this patch, the size of a transcoded file was returned as the
size of the raw file. In the case of at least one file format (FLAC) the
raw file is much  larger than the transcoded response.

With this path, subdaap requests the estimated content size from
subsonic. This size is then used in the response that is returned to the
client while streaming.